pyproject.toml
[project]
name = "dvre"
version = "0.1.0"
description = "DaVinci Resolve Video Editor - Server for creating timelines via DaVinci Resolve scripting API"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi",
    "uvicorn",
    "pydantic",
    "psutil",
    "python-dotenv",
    "fusionscript-stubs",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["dvre"]
